LayoutTests/security/contentSecurityPolicy/worker-inherits-blocks-xhr-expected.txt

CONSOLE MESSAGE: Refused to connect to non-existent-file because it does not appear in the connect-src directive of the Content Security Policy.
This tests that the Content Security Policy (CSP) of the owner document (this page) blocks a file-URL Web Worker from making an XHR request because the parent's CSP contains "connect-src 'none'"

PASS threw exception Error: NetworkError: DOM Exception 19.
